+++ VocaDbWeb/Scripts/App.tsx
@@ -1,4 +1,3 @@
-import { AppProvider } from '@/AppContext';
 import AppRoutes from '@/AppRoutes';
 import Container from '@/Bootstrap/Container';
 import { AboutDisclaimer } from '@/Components/Shared/Partials/AboutDisclaimer';
@@ -34,7 +33,7 @@ const AppContainer = (): React.ReactElement => {
 
 const App = (): React.ReactElement => {
 	return (
-		<AppProvider>
+		<>
 			<ScrollToTop />
 			<Header />
 			<div css={{ display: 'flex' }}>
@@ -43,7 +42,7 @@ const App = (): React.ReactElement => {
 			</div>
 			<Toaster containerStyle={{ top: '10vh' }} gutter={0} />
 			<VdbPlayer />
-		</AppProvider>
+		</>
 	);
 };
