+++ VocaDbWeb/Scripts/Components/Shared/Partials/Comment/CommentBodyLarge.tsx
@@ -1,10 +1,10 @@
-import { useMutedUsers } from '@/AppContext';
 import { MomentJsTimeAgo } from '@/Components/KnockoutExtensions/MomentJsTimeAgo';
 import { FormatMarkdown } from '@/Components/Shared/Partials/Html/FormatMarkdown';
 import { ProfileIcon } from '@/Components/Shared/Partials/User/ProfileIcon';
 import { UserLink } from '@/Components/Shared/Partials/User/UserLink';
 import { CommentContract } from '@/DataContracts/CommentContract';
 import { LoginManager } from '@/Models/LoginManager';
+import { useMutedUsers } from '@/MutedUsersContext';
 import { EntryUrlMapper } from '@/Shared/EntryUrlMapper';
 import { observer } from 'mobx-react-lite';
 import React from 'react';
